% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan.R
\name{handle_error}
\alias{handle_error}
\title{Get and free the error message stored at the C++ pointer}
\usage{
handle_error(lib_name, err_msg, err_ptr, function_name)
}
\description{
Get and free the error message stored at the C++ pointer
}
\keyword{internal}
