% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{set_bridgestan_path}
\alias{set_bridgestan_path}
\title{Function \code{set_bridgestan_path()}}
\usage{
set_bridgestan_path(path)
}
\description{
Set the path to BridgeStan.
}
\details{
This should point to the top-level folder of the repository.
}
