% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{verify_bridgestan_path}
\alias{verify_bridgestan_path}
\title{Get the path to BridgeStan.}
\usage{
verify_bridgestan_path(path)
}
\description{
By default this is set to the value of the environment
variable \code{BRIDGESTAN}.
}
\details{
If there is no path set, this function will download
a matching version of BridgeStan to a folder called
\code{.bridgestan} in the user's home directory.

See also \code{set_bridgestan_path}
}
