% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan.R
\name{StanRNG}
\alias{StanRNG}
\title{StanRNG}
\description{
StanRNG

StanRNG
}
\details{
RNG object for use with \code{StanModel$param_constrain()}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rng}}{The pointer to the RNG object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanRNG-new}{\code{StanRNG$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanRNG-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanRNG-new}{}}}
\subsection{Method \code{new()}}{
Create a StanRng
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanRNG$new(lib_name, rng_seed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lib_name}}{The name of the Stan dynamic library.}

\item{\code{rng_seed}}{The seed for the RNG.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new StanRNG.
}
}
}
